requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
flasgger==0.9.7.1
psycopg2-binary==2.9.7
Werkzeug==2.3.7
setuptools==68.2.2
pytest==7.4.0
requests-mock==1.11.0